Create the CIFTI-2 XML metadata for a set of axes.

Build the CIFTI-2 XML document (the part of a CIFTI-2 file that describes what the data matrix contains) from a set of axes. This is the inverse of the XML parsing that read.cifti.header performs: pass the result of this function to cifti.parse.xml or to read.cifti.header (after storing it in a file) and you get back the axes you put in.

Usage

cifti.header.from.axes(axes, metadata = NULL)

Arguments

axes

list of axes, one per matrix dimension, as created by the cifti.axis.*() functions or by cifti.axis.from.template. If the list is named, the names have to be the matrix dimensions ('0' and '1'), otherwise the axes are assigned to the dimensions in the order in which they are given.

metadata

named character vector or named list or NULL, the metadata of the matrix (e.g. the provenance information), as name/value pairs.

Value

character string, the CIFTI-2 XML document.

See also

Examples

axis_series <- cifti.axis.series(4L, start = 0, step = 2.5)
axis_brain <- cifti.axis.brain.models(list(
  cifti.brain.model.surface("lh", 10L),
  cifti.brain.model.surface("rh", 12L)))
xml <- cifti.header.from.axes(list(axis_series, axis_brain))
cat(substr(xml, 1, 200))
#> <?xml version="1.0" encoding="UTF-8"?>
#> <CIFTI Version="2">
#>   <Matrix>
#>     <MatrixIndicesMap AppliesToMatrixDimension="0" IndicesMapToDataType="CIFTI_INDEX_TYPE_SERIES" NumberOfSeriesPoints="4" SeriesS