Create a NIFTI v1 header from the header information contained in an fs.volume instance.

Create a NIFTI v1 header from the header information contained in an fs.volume instance.

Usage

nii1header.for.mgh(mgh, endian = "little")

Arguments

mgh

an fs.volume instance, or a string. If a string, it is interpreted as a filepath to a volume file (NIFTI, MGH or MGZ) that will be loaded with read.fs.volume.

endian

character string, the endianness to use. Either 'little' or 'big'. Defaults to 'little'.

Value

a NIFTI v1 header structure (see ni1header.template). Note that the header may or may not contain full RAS information, depending on whether the source fs.volume contained such information or not. If the MGH header does not have valid RAS information, the qform and sform codes will be set to 0 (unknown).

Note

This is intended to be used with write.nifti1, which allows users to convert MGH/MGZ data to NIFTI files.

See also