Read a FreeSurfer ASCII surface file.

Read a FreeSurfer ASCII surface file.

Usage

# S3 method for class 'fsascii'
read_nisurfacefile(filepath, ...)

Arguments

filepath

character string, the full path to the input surface file.

...

parameters passed to read.fs.surface.asc.

Value

an instance of fs.surface, read from the file. See read.fs.surface for details. If none of the reader methods succeed, an error is raised.