Adapt spatial transformation matrix for 1-based indices.

Adapt spatial transformation matrix for 1-based indices.

sm0to1(tf_matrix)

Arguments

tf_matrix
4x4 numerical matrix, the input spatial transformation matrix, suitable for 0-based indices. Typically this is a vox2ras matrix obtained from functions like mghheader.vox2ras.

Value

4x4 numerical matrix, adapted spatial transformation matrix, suitable for 1-based indices.

See also