Write file in FreeSurfer weight ASCII format

Write vertex-wise brain data for a set of vertices to an ASCII file in weight format. This format is also known as paint format or simply as w format.

Usage

write.fs.weight.asc(filepath, vertex_indices, values)

Arguments

filepath,

string. Full path to the output ASCII weight file.

vertex_indices

vector of integers, the vertex indices. Must be one-based (R-style). This function will substract 1, as they need to be stored zero-based in the file.

values

vector of floats. The brain morphometry data to write, one value per vertex.

See also